Moved classes ImportController
Moved static classes form ImportController to file ColumnListItem, FileListItem and GroupListItem
